Move screen components from ui to screen folder
